Move use-icon hook to metabase/common/hooks
