+++ frontend/src/metabase/common/components/Pickers/EntityPicker/utils.ts
@@ -2,7 +2,7 @@ import { useCallback } from "react";
 import _ from "underscore";
 
 import type { IconData } from "metabase/common/utils/icon";
-import { useGetIcon } from "metabase/hooks/use-icon";
+import { useGetIcon } from "metabase/common/hooks/use-icon";
 import type { ColorName } from "metabase/ui/colors/types";
 import type {
   CollectionItemModel,
