+++ frontend/src/metabase/common/hooks/use-icon.ts
