+++ frontend/src/metabase/common/hooks/use-icon.unit.spec.ts
