+++ frontend/src/metabase/query_builder/components/view/sidebars/QuestionInfoSidebar/components/hooks.ts
@@ -1,6 +1,6 @@
 import { useCallback, useState } from "react";
 
-import { useGetIcon } from "metabase/hooks/use-icon";
+import { useGetIcon } from "metabase/common/hooks/use-icon";
 import { getUrl } from "metabase/querying/notebook/components/NotebookDataPicker/utils";
 import * as Lib from "metabase-lib";
 import type Question from "metabase-lib/v1/Question";
