+++ frontend/src/metabase/query_builder/components/view/sidebars/QuestionInfoSidebar/components/hooks.unit.spec.ts
@@ -11,7 +11,7 @@ import {
 
 import { useGetJoinedTablesWithIcons } from "./hooks";
 
-jest.mock("metabase/hooks/use-icon", () => {
+jest.mock("metabase/common/hooks/use-icon", () => {
   const { modelIconMap } = jest.requireActual("metabase/common/utils/icon");
   return {
     useGetIcon: () =>
